+++ js/main.js
@@ -1,5 +1,5 @@
-import './util.js';
-import './data.js';
+import {comments, names} from './data.js';
+import {createDescriptions, checkLengthComment} from './util.js';
 
-
-
+createDescriptions(comments, names);
+checkLengthComment('sjefhksjfh', 90);
